requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
google-cloud-speech>=2.26.0
google-genai>=1.0.0
google-cloud-texttospeech>=2.16.0
websockets>=12.0
pyaudio>=0.2.14
python-dotenv>=1.0.0
pytest>=8.0
pytest-asyncio>=0.23
